Beginning Main Webpage
started our new index.html with beginnings of basic website structure (navbar and content/profile section).
I suck with flexbox but I was trying to make the content blocks (posts) in one column, and the profile section in the column to the left of it, similar to facebook or instagram.